image_translator_merged/return/requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
python-dotenv>=0.21.0
redis>=4.5.0
